// frontend/src/api/model/judgeResultResponse.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export interface JudgeResultResponse {
  id?: string;
  submissionId?: string;
  testCaseId?: string;
  status?: string;
  executionTimeMs?: number;
  memoryKb?: number;
  createdAt?: string;
}